RCT of Indomethacin for Prevention of Post-ERCP Pancreatitis

Results of a randomized, placebo-controlled, prospective 2-arm trial of rectal indomethacin (100 mg) vs. placebo prevent post-ERCP pancreatitis in 602 participants, as reported by Elmunzer, Higgins, et al. in 2012 in the New England Journal of Medicine.
ERCP, or endoscopic retrograde cholangio-pancreatogram, is a procedure performed by threading an endoscope through the mouth to the opening in the duodenum where bile and pancreatic digestive juices are released into the intestine. ERCP is helpful for treating blockages of flow of bile (gallstones, cancer), or diagnosing cancers of the pancreas, but has a high rate of complications (15-25%).
The occurrence of post-ERCP pancreatitis is a common and feared complication, as pancreatitis can result in multisystem organ failure and death, and can occur in ~ 16% of ERCP procedures.
The inflammatory cytokine storm that can result from this procedural complication can be quite severe. Several small randomized trials suggested that anti-inflammatory NSAID therapies at the time of ERCP could reduce the rate of this complication, but all were rather small single-center studies, and were not sufficiently convincing to change practice. /cr Elmunzer, Higgins, and colleagues performed a meta-analysis of these small trials, which suggested that this was a significant effect, and that indomethacin could result in a 64% reduction in post-ERCP pancreatitis.
The investigators took this as a possible over-estimate of the effect (due to publication bias), and designed a multicenter RCT of a planned 948 patients to see a reduction of 50% from a placebo rate of 10% to an indomethacin rate of 5%. Two interim analyses were performed, after 400 and 600 patients were enrolled, using an alpha spending function. The Data and Safety Monitoring Board stopped the study after 602 participants were enrolled because of the significantly positive effect of indomethacin, which reduced post-ERCP pancreatitis from 16% in the placebo group to 9% in the indomethacin group.
You can find the manuscript at Indomethacin to Prevent Post-ERCP Pancreatitis.

indo_rct

Format

A data frame with 602 rows and 25 variables

subject

subject id, first integer indicates center, integer, range:1001-4003

age

age in years, numeric, range: 19-90

risk

risk score, numeric, range: 1-5.5

gender

male or female, factor, ordered: FALSE, levels: 1. 1_female, 2. 2_male

outcome

outcome of post-ercp pancreatitis, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

group

treatment arm, factor, ordered: FALSE, levels: 1. 1_placebo, 2. 2_indomethacin

sod

sphincter of oddi dysfunction was present, a risk factor favoring post-ERCP pancreatitis, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

pep

previous post-ERCP pancreatitis (PEP), a risk factor for future PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

recpancreatitis

Recurrent Pancreatitis, a risk factor for future PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

psphinct

a Pancreatic Sphincterotomy was performed, a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

precut

a sphincter pre-cut was needed to enter the papilla, a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

diffcann

Cannulation of the papilla was difficult, a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

pancinj

Contrast was injected into the pancreas dring the procedure, a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

somsod

Sphincter of oddi manometry was performed during the procedure for SOD, a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

pdstent

A pancreatic duct stent was placed at the end of the procedure per the judgement of the endoscopist (more often in high-risk cases), a potential protective effect against PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

trainee

A trainee participated in the ERCP, which could be a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

bilsphinct

A billiary sphincterotomy was performed, which could be a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

ampullectomy

An Ampullectomy was performed for dysplasia or cancer, which could be a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

acinarization

The pancreas appeared to have acinarization on imaging, which could be a risk factor for PEP, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

bilstent

A biliary stent was placed, which could be protective. factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

choledocho

Choledocholithiasis (gallstones blocking the biliary duct) was present, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

cancer

Cancer of the biliary duct or pancreas was found, factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

pbrush

Brushings were taken from the pancreatic duct, a possible risk factor favoring post-ERCP pancreatitis factor, ordered: FALSE, levels: 1. 1_yes, 2. 2_no

gibleed

A gastrointestinal bleed occurred (which could be a complication of indomethacin therapy), factor, ordered: FALSE, levels: 1. 0_no, 2. 1_yes

mspep

NA, factor, ordered: FALSE, levels: 1. 0_no, 2. 1_yes

Source

This data set is sourced from the authors of the 2012 manuscript in the New England Journal of Medicine, entitled, A Randomized Trial of Rectal Indomethacin to Prevent Post-ERCP Pancreatitis, pages 1414-1422 volume 366, in the April 12, 2012 edition, authored by the Elmunzer, BJ, Higgins PDR, et al.

You can find the manuscript at Indomethacin to Prevent Post-ERCP Pancreatitis.